+++ packages/vscode-container-client/src/index.ts
@@ -7,6 +7,7 @@ export * from './clients/DockerClient/DockerClient';
 export * from './clients/DockerComposeClient/DockerComposeClient';
 export * from './clients/PodmanClient/PodmanClient';
 export * from './clients/PodmanComposeClient/PodmanComposeClient';
+export * from './clients/WslcClient/WslcClient';
 export * from './commandRunners/shellStream';
 export * from './commandRunners/wslStream';
 export * from './contracts/CommandRunner';